fix diem thi + nhhap diem + lop hoc ttth

+++ projects/admin/src/app/components/modal/ttth/modal-ttth-quanlidiemthi/modal-ttth-quanlidiemthi.component.ts
@@ -3,7 +3,7 @@ import { ModalService } from '../../../../services/modal.service';
 import { DiemthiService } from '../../../../services/ttth/diemthi.service';
 import { ttthDiemThi } from '../../../../../models/ttthDiemThi';
 import { ToastrService } from 'ngx-toastr';
-
+import { getCookie } from '../../../../../../../common/helper';
 @Component({
   selector: 'app-modal-ttth-quanlidiemthi',
   templateUrl: './modal-ttth-quanlidiemthi.component.html',
@@ -12,6 +12,7 @@ import { ToastrService } from 'ngx-toastr';
 export class ModalTtthQuanlidiemthiComponent implements OnInit {
   constructor(private modalService: ModalService,private DiemthiService: DiemthiService ,private toastr: ToastrService) { }
   DiemThi: ttthDiemThi[];
+  private _username: any = getCookie('displayName');
   ngOnInit(): void {
     this.getdanhsach();
   }
@@ -20,7 +21,7 @@ export class ModalTtthQuanlidiemthiComponent implements OnInit {
   }
 
   getdanhsach(): void {
-    this.DiemthiService.get().subscribe((data) => {this.DiemThi = data});
+    this.DiemthiService.get().subscribe((data) => {this.DiemThi = data;setTimeout(() => {}, 500);});
   }
   reset():void{
     this.selectedItem=null;
@@ -32,22 +33,28 @@ export class ModalTtthQuanlidiemthiComponent implements OnInit {
   }
   update(DiemThi: ttthDiemThi):void {
     DiemThi.updated_at= new Date;
+    DiemThi.nguoisua= this._username;
     this.DiemthiService.update(DiemThi)
     .subscribe(data => {
       this.DiemThi.push(data);
+      this.getdanhsach();
     });
     this.toastr.success('Sửa thành công');
   }
   //delete
   delete(DiemThi: ttthDiemThi):void {
-    this.DiemthiService.delete(DiemThi)
-    .subscribe(data => {
-      this.DiemThi.push(data);
-      setTimeout(() => {}, 0);
-    });
-    this.getdanhsach();
-    // window.location.reload();
-    this.toastr.success('Xóa thành công');
+    var comfirmDel = confirm('Bạn có chắc chắn muốn xóa');
+    if(comfirmDel==true){
+      DiemThi.nguoisua= this._username;
+      DiemThi.updated_at= new Date;
+
+      this.DiemthiService.delete(DiemThi)
+      .subscribe(data => {
+        this.DiemThi.push(data);
+        this.getdanhsach();
+      });
+      this.toastr.success('Xóa thành công');
+    }
   }
 
 }
